minio_container/requirements.txt
minio
fastapi
uvicorn
boto3
python-multipart
python-dotenv
pydantic-settings
